Fix: 마이프로필 팔로우 버튼 기능 #137
