+++ src/pages/Profile/Follow.jsx
@@ -9,30 +9,22 @@ import {
 } from '../../hooks/queries/useProfile';
 import { userInfoAtom } from '../../library/atom';
 import { useRecoilValue } from 'recoil';
-
+import useFollowUser from '../../hooks/queries/useFollow';
 export default function Follow() {
   const navigate = useNavigate();
   const location = useLocation();
   const UserId = useRecoilValue(userInfoAtom).id;
   const [activeList, setActiveList] = useState('followers');
   const [refreshData, setRefreshData] = useState(false);
-
-  const {
-    data: followers,
-    isLoading: followingLoading,
-    refetch: refetchFollowers,
-  } = useGetFollower(UserId);
-  const {
-    data: followings,
-    isLoading: followerLoading,
-    refetch: refetchFollowings,
-  } = useGetFollowing(UserId);
+  const [users, setUsers] = useState();
+  const { data: followers, isLoading: followingLoading } =
+    useGetFollower(UserId);
+  const { data: followings, isLoading: followerLoading } =
+    useGetFollowing(UserId);
 
   useEffect(() => {
     const shouldRefetch = refreshData || location.state?.type;
     if (shouldRefetch) {
-      refetchFollowers();
-      refetchFollowings();
       setRefreshData(false);
     }
 
@@ -42,24 +34,46 @@ export default function Follow() {
     ) {
       setActiveList(location.state.type);
     }
-  }, [
-    refreshData,
-    location.state,
-    refetchFollowers,
-    refetchFollowings,
-    UserId,
-  ]);
+  }, [refreshData, location.state, UserId]);
 
-  const handleFollowClick = (userData) => {
-    setRefreshData(true);
+  // const handleFollowClick = async (user) => {
+  //   console.log(user);
+  // };
+
+  const { followUser } = useFollowUser(); // 커스텀 훅 사용
+
+  const handleFollowClick = (user) => {
+    const isUnfollowing = user.isFollowing;
+    followUser(user.id, isUnfollowing, {
+      onSuccess: () => {
+        // 성공적으로 상태 변경 후 사용자 목록 업데이트
+        if (isUnfollowing) {
+          // 언팔로우하는 경우: 사용자를 목록에서 제거
+          setUsers((currentUsers) =>
+            currentUsers.filter((u) => u.id !== user.id),
+          );
+        } else {
+          // 팔로우하는 경우: 사용자의 isFollowing 상태를 업데이트
+          setUsers((currentUsers) =>
+            currentUsers.map((u) =>
+              u.id === user.id ? { ...u, isFollowing: true } : u,
+            ),
+          );
+        }
+      },
+    });
   };
 
-  const renderUserList = (data, listType) => {
+  const renderUserList = (data, listType, onFollowClick) => {
     if (!Array.isArray(data) || data.length === 0) {
       return <p id='FollowNone'>앗! 아직 비어있어요</p>;
     }
     return (
-      <FollowUserList users={data.map(transformUserData)} listType={listType} />
+      <FollowUserList
+        users={data.map(transformUserData)}
+        listType={listType}
+        onFollowClick={onFollowClick}
+      />
     );
   };
 
@@ -90,8 +104,10 @@ export default function Follow() {
           {followings?.length ?? 0} 팔로잉
         </ListToggleButton>
       </ListToggleButtonWrap>
-      {activeList === 'followers' && renderUserList(followers, 'followers')}
-      {activeList === 'followings' && renderUserList(followings, 'followings')}
+      {activeList === 'followers' &&
+        renderUserList(followers, 'followers', handleFollowClick)}
+      {activeList === 'followings' &&
+        renderUserList(followings, 'followings', handleFollowClick)}
     </FollowWrap>
   );
 }
